+++ gui/src/helpers/core/__init__.py
@@ -8,6 +8,7 @@ from .recommendation_worker import RecommendationWorker
 from .sampler_worker import SamplerWorker
 from .search_worker import SearchWorker
 from .similarity_scan_worker import SimilarityScanWorker
+from .video_export_worker import ScrollVideoExportWorker
 from .wallpaper_worker import WallpaperWorker
 
 __all__ = [
@@ -19,6 +20,7 @@ __all__ = [
     "QueueExecutionWorker",
     "RecommendationWorker",
     "SamplerWorker",
+    "ScrollVideoExportWorker",
     "SearchWorker",
     "SimilarityScanWorker",
     "WallpaperWorker",